Fix to the 404 issue #159.
What I have found is that nginx does not give us the names in lowercase, it is in fact our web browsers which convert hostnames to lowercase. Nginx is just passing that lowercase hostname on to server.php. As Windows is incase-sensitive, we can strtolower() on both variables to ensure that this works for everyone.